refactor(WifiListItem): move WifiListItem to shared utility file for reuse
- Relocated WifiListItem composable to `com.greybox.projectmesh.extension.WifiListItem`
- Removed old definitions from NetworkScreen and replaced them with imports of the new utility
- Updated ChatNodeListScreen and NetworkScreen to reference the shared component
- Ensures a single source of UI truth for node listing and improves code maintainability
